Moved code in main.py to game.py to create segregation for distinct sections of code. Added interface.py which will be used to create the GUI for the game and simulations.
