+++ interface.py
@@ -1,0 +1,5 @@
+import tkinter
+
+# TODO: Create interface
+# TODO: Hook interface into game
+pass
